Move types around to make types.ts pure
